amd-kria: avoid truncating the FRU board area offset

+++ plugins/amd-kria/fu-amd-kria-som-eeprom.c
@@ -35,7 +35,7 @@ fu_amd_kria_som_eeprom_parse(FuFirmware *firmware,
 	FuAmdKriaSomEeprom *self = FU_AMD_KRIA_SOM_EEPROM(firmware);
 	guint8 str_len = 0;
 	guint8 str_offset = FU_STRUCT_BOARD_INFO_OFFSET_MANUFACTURER_LEN;
-	guint8 board_offset;
+	gsize board_offset;
 	const guint8 *buf;
 	gsize bufsz = 0;
 	g_autoptr(FuStructIpmiCommon) st_common = NULL;
